+++ src/genui/ui_widgets/window_mixins/generation_command.py
@@ -34,7 +34,7 @@ class GenerationCommandMixin:
 
     def handle_generate(self):
         if not self._validate_data_for_generation_method():
-            self.show_modal_dialog()
+            self.show_error_modal_dialog()
             return
 
         self.button_generate.setDisabled(True)
@@ -43,7 +43,7 @@ class GenerationCommandMixin:
         try:
             self._generate_method()
         except ValueError as e:
-            self.show_modal_dialog(str(e))
+            self.show_error_modal_dialog(str(e))
             self.button_generate.setDisabled(False)
             self.button_interrupt.setDisabled(True)
 
@@ -53,16 +53,16 @@ class GenerationCommandMixin:
         # FIXME: `self.model_path` can be changed. Need using prompt.
         interrupt(self.model_path)
 
-    def show_modal_dialog(self, err_data: str | None = None):
+    def show_error_modal_dialog(self, err_data: str | None = None):
         """Show a modal dialog with an error message."""
 
         msg = QtWidgets.QMessageBox()
-        msg.setWindowTitle("Data error")
+        msg.setWindowTitle("Error")
 
         if err_data:
-            msg.setText(f"Data for generation not valid! {err_data}.")
+            msg.setText(err_data)
         else:
-            msg.setText("Data for generation not valid! Try check model is chose.")
+            msg.setText("Data for generation not valid! Try check a model is chosen.")
 
         msg.setWindowModality(Qt.WindowModality.ApplicationModal)